requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.12.1email-validator==2.1.0boto3==1.29.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.13.0
celery==5.3.4
redis==5.0.1
pymongo==4.6.0
motor==3.3.2
